+++ writer/arrow.go
@@ -46,6 +46,9 @@ func NewArrowWriter(arrowSchema *arrow.Schema, pfile source.ParquetFileWriter, o
 	if err = res.initBloomFilters(); err != nil {
 		return nil, fmt.Errorf("init bloom filters: %w", err)
 	}
+	if err = res.validateEncryptionColumnKeys(); err != nil {
+		return nil, err
+	}
 
 	res.stopped = false
 	return res, nil
