+++ writer/csv.go
@@ -39,6 +39,9 @@ func NewCSVWriter(md []string, pfile source.ParquetFileWriter, opts ...WriterOpt
 	if err = res.initBloomFilters(); err != nil {
 		return nil, fmt.Errorf("init bloom filters: %w", err)
 	}
+	if err = res.validateEncryptionColumnKeys(); err != nil {
+		return nil, err
+	}
 
 	res.stopped = false
 	return res, nil
