+++ writer/json_test.go
@@ -341,3 +341,39 @@ func TestJSONWriter(t *testing.T) {
 		}
 	})
 }
+
+func TestJSONWriterValidatesEncryptionColumnKeys(t *testing.T) {
+	t.Parallel()
+
+	jsonSchema := `{
+		"Tag": "name=parquet_go_root",
+		"Fields": [
+			{"Tag": "name=id, type=INT32"},
+			{"Tag": "name=name, type=BYTE_ARRAY, convertedtype=UTF8"}
+		]
+	}`
+
+	t.Run("typo in column key path is rejected", func(t *testing.T) {
+		t.Parallel()
+		var buf bytes.Buffer
+		_, err := NewJSONWriter(
+			jsonSchema,
+			writerfile.NewWriterFile(&buf),
+			WithFooterKey([]byte("0123456789abcdef")),
+			WithColumnKey("nmae", []byte("abcdef0123456789")),
+		)
+		require.ErrorContains(t, err, "nmae")
+	})
+
+	t.Run("valid column key path is accepted", func(t *testing.T) {
+		t.Parallel()
+		var buf bytes.Buffer
+		_, err := NewJSONWriter(
+			jsonSchema,
+			writerfile.NewWriterFile(&buf),
+			WithFooterKey([]byte("0123456789abcdef")),
+			WithColumnKey("name", []byte("abcdef0123456789")),
+		)
+		require.NoError(t, err)
+	})
+}
